requirements.txt
Django>=2.1.0,<2.2.0
djangorestframework>=3.8.2,<3.9.0
psycopg2>=2.7.5,<2.8.0
flake8>=3.6.0,<3.7.0
django-cors-headers>=3.4.0